WIP: migrate types, functionality to packages
