+++ asset/src/elasticsearch_reader/interfaces.ts
@@ -9,70 +9,10 @@ import moment from 'moment';
 import { DataTypeConfig } from '@terascope/types';
 import WindowState from '../elasticsearch_reader_api/window-state';
 import { IDType, WildCardQuery } from '../id_reader/interfaces';
+import { DateSegments, ParsedInterval } from '../../../packages/elasticsearch-asset-apis/interfaces';
 
 export interface ESReaderConfig extends ESReaderOptions, OpConfig {
     api_name: string;
-}
-
-export interface ESReaderOptions {
-    index: string;
-    id_field_name?: string;
-    size: number;
-    start: null | string;
-    end: null | string;
-    interval: string;
-    preserve_id: boolean;
-    date_field_name: string;
-    query?: string;
-    fields: null | string[];
-    delay: string;
-    subslice_by_key: boolean;
-    subslice_key_threshold: number;
-    key_type: IDType;
-    type: string | null;
-    time_resolution: moment.unitOfTime.Base;
-    geo_field?: string;
-    geo_box_top_left?: string;
-    geo_box_bottom_right?: string;
-    geo_point?: string;
-    geo_distance?: string;
-    geo_sort_point?: string;
-    geo_sort_order?: string;
-    geo_sort_unit?: string;
-    connection: string;
-    starting_key_depth: number;
-    use_data_frames?: boolean;
-    type_config?: DataTypeConfig
-}
-
-export interface InputDateSegments {
-    start: moment.Moment | string | Date;
-    limit: moment.Moment | string | Date;
-}
-
-export interface DateSegments {
-    start: moment.Moment;
-    limit: moment.Moment;
-}
-
-export interface SlicerDateConfig extends DateSegments {
-    end: moment.Moment;
-    holes?: DateConfig[];
-}
-
-export interface StartPointConfig {
-    dates: DateSegments;
-    id: number;
-    numOfSlicers: number;
-    interval: ParsedInterval;
-    recoveryData?: SlicerDateResults[];
-}
-
-export type ParsedInterval = [number, moment.unitOfTime.Base];
-
-export interface DateConfig {
-    start: string | moment.Moment;
-    end: string | moment.Moment;
 }
 
 export interface SlicerArgs {
@@ -100,27 +40,4 @@ export interface SlicerDateResults {
     key?: string;
 }
 
-export interface ApiConfig extends ESReaderConfig {
-    endpoint: string;
-    token: string;
-    timeout: number;
-    full_response?: boolean;
-}
-
 export type ESDateConfig = ESReaderConfig | ApiConfig;
-
-export interface DetermineSliceResults {
-    start: moment.Moment;
-    end: moment.Moment;
-    count: number;
-    key?: string;
-}
-
-export interface IDReaderSlice {
-    start?: string;
-    end?: string;
-    count: number;
-    holes?: DateConfig[];
-    wildcard?: WildCardQuery;
-    key?: string;
-}
