+++ package.json
@@ -9,8 +9,8 @@
     ],
     "scripts": {
         "asset:build": "yarn && yarn run build",
-        "build": "yarn workspaces run build",
-        "build:watch": "yarn build --watch",
+        "build": "yarn wsrun --stages build",
+        "build:watch": "yarn wsrun --stages --parallel build --watch",
         "lint": "eslint --ignore-path .gitignore --ext .js,.ts .",
         "lint:fix": "yarn lint --fix",
         "test:all": "yarn workspaces run test",
@@ -57,6 +57,7 @@
         "teraslice-test-harness": "^0.27.0",
         "ts-jest": "^26.4.4",
         "ts-node": "^9.1.1",
-        "typescript": "^4.1.3"
+        "typescript": "^4.1.3",
+        "wsrun": "^5.2.4"
     }
 }
