+++ packages/elasticsearch-asset-apis/src/elasticsearch_id_slicer/helpers.ts
@@ -1,6 +1,6 @@
 /* eslint-disable no-useless-escape */
 
-import { IDType } from '../../id_reader/interfaces';
+import { IDType } from '../interfaces';
 
 export const base64url = ['a', 'b', 'c', 'd', 'e', 'f', 'g', 'h', 'i', 'j', 'k', 'l', 'm', 'n', 'o', 'p', 'q', 'r', 's', 't', 'u', 'v', 'w',
     'x', 'y', 'z', 'A', 'B', 'C', 'D', 'E', 'F', 'G', 'H', 'I', 'J', 'K', 'L', 'M', 'N', 'O', 'P', 'Q', 'R', 'S', 'T', 'U', 'V', 'W', 'X',
