+++ packages/elasticsearch-asset-apis/src/index.ts
@@ -1,0 +1,5 @@
+export * from './interfaces';
+export * from './window-state';
+export * from './spaces-api';
+export * from './elasticsearch-api';
+export * from './base-api';
