+++ packages/elasticsearch-asset-apis/src/window-state.ts
