+++ packages/terafoundation_elasticsearch_connector/jest.config.js
@@ -1,0 +1,22 @@
+'use strict';
+
+module.exports = {
+    verbose: true,
+    testEnvironment: 'node',
+    setupFilesAfterEnv: ['jest-extended'],
+    collectCoverage: true,
+    coverageReporters: ['json', 'lcov', 'text', 'html'],
+    coverageDirectory: 'coverage',
+    testMatch: [
+        '<rootDir>/test/**/*-spec.{ts,js}',
+        '<rootDir>/test/*-spec.{ts,js}',
+    ],
+    preset: 'ts-jest',
+    globals: {
+        'ts-jest': {
+            tsConfig: './tsconfig.json',
+            diagnostics: true,
+            pretty: true,
+        }
+    }
+};
