+++ packages/terafoundation_elasticsearch_connector/package.json
@@ -1,0 +1,28 @@
+{
+    "name": "@terascope/terafoundation_elasticsearch_connector",
+    "version": "0.1.0",
+    "description": "Elasticsearch connector for terafoundation",
+    "publishConfig": {
+        "access": "public"
+    },
+    "srcMain": "src/index.ts",
+    "main": "dist/index.js",
+    "typings": "dist/index.d.ts",
+    "scripts": {
+        "prepare": "yarn build",
+        "build": "tsc --project tsconfig.json",
+        "build:watch": "yarn build --watch",
+        "test": "jest",
+        "test:watch": "jest --coverage=false --notify --watch --onlyChanged",
+        "test:debug": "env DEBUG='*teraslice*' jest --detectOpenHandles --coverage=false --runInBand"
+    },
+    "homepage": "https://github.com/terascope/terafoundation_elasticsearch_connector",
+    "repository": "git@github.com:terascope/terafoundation_elasticsearch_connector.git",
+    "author": "Terascope, LLC <info@terascope.io>",
+    "license": "MIT",
+    "dependencies": {},
+    "devDependencies": {},
+    "engines": {
+        "node": ">=10.16.0"
+    }
+}
