+++ packages/terafoundation_elasticsearch_connector/src/index.ts
