+++ packages/terafoundation_elasticsearch_connector/src/schema.ts
