+++ packages/terafoundation_elasticsearch_connector/tsconfig.json
@@ -1,0 +1,14 @@
+{
+    "extends": "../../tsconfig",
+    "compilerOptions": {
+        "baseUrl": ".",
+        "rootDir": "src",
+        "outDir": "dist",
+        "declaration": true,
+        "sourceMap": false,
+        "inlineSourceMap": true
+    },
+    "include": [
+        "src/**/*"
+    ]
+}
